refactor(nexp): eventhandler as injectable dependency
